requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.13
    # via vipbot (pyproject.toml)
aiorun==2025.1.1
    # via vipbot (pyproject.toml)
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via starlette
attrs==25.1.0
    # via aiohttp
beanie==1.29.0
    # via vipbot (pyproject.toml)
click==8.1.8
    # via
    #   beanie
    #   uvicorn
colorlog==6.9.0
    # via vipbot (pyproject.toml)
dnspython==2.7.0
    # via pymongo
fastapi==0.115.11
    # via vipbot (pyproject.toml)
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   yarl
lazy-model==0.2.0
    # via beanie
motor==3.7.0
    # via beanie
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
pyaes==1.6.1
    # via pyrofork
pydantic==2.10.6
    # via
    #   vipbot (pyproject.toml)
    #   beanie
    #   fastapi
    #   lazy-model
pydantic-core==2.27.2
    # via pydantic
pymediainfo-pyrofork==6.0.2
    # via pyrofork
pymongo==4.11.1
    # via motor
pyrofork==2.3.58
    # via vipbot (pyproject.toml)
pysocks==1.7.1
    # via pyrofork
python-dotenv==1.0.1
    # via vipbot (pyproject.toml)
sniffio==1.3.1
    # via anyio
starlette==0.46.0
    # via fastapi
tgcrypto-pyrofork==1.2.7
    # via vipbot (pyproject.toml)
toml==0.10.2
    # via beanie
typing-extensions==4.12.2
    # via
    #   anyio
    #   beanie
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn==0.34.0
    # via vipbot (pyproject.toml)
yarl==1.18.3
    # via aiohttp
